Refactor time parser utilities
This commit refactors away the time parser utilities that were present
inItially in the Task class.

Why I did so was because it did not make sense for a Task class to contain
date parsing utilities.
